+++ sample-compose-fragments/src/main/java/uk/gov/hmrc/sample_compose_fragments/data/model/AtomItem.kt
@@ -13,20 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.hmrc.sample_compose_fragments.presentation.fragment
+package uk.gov.hmrc.sample_compose_fragments.data.model
 
-import android.os.Bundle
-import android.view.View
-import androidx.fragment.app.Fragment
-import uk.gov.hmrc.sample_compose_components.R
-import uk.gov.hmrc.sample_compose_components.databinding.FragmentAtomsBinding
+import androidx.annotation.StringRes
+import uk.gov.hmrc.sample_compose_fragments.presentation.screens.ComponentItem
 
-class AtomsFragment : Fragment(R.layout.fragment_atoms) {
-
-    private lateinit var binding: FragmentAtomsBinding
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding = FragmentAtomsBinding.bind(view)
-    }
-}
+class AtomItem(val id: Int, @StringRes override val titleRes: Int): ComponentItem