+++ sample-compose-fragments/src/main/java/uk/gov/hmrc/sample_compose_fragments/presentation/fragment/colors/ColorsFragment.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.hmrc.sample_compose_fragments.presentation.fragment
+package uk.gov.hmrc.sample_compose_fragments.presentation.fragment.colors
 
 import android.os.Bundle
 import android.view.View