+++ CHANGELOG.md
@@ -20,6 +20,10 @@ Allowed headings:
 
 * `EditableListView` changed to use `EditableListItemViewState`.
 
+### Additions
+
+* Added new `MiniAdvertCardView` component.
+
 ## [4.4.1] - 2023-02-03
 
 ### Changed